+++ modules/@apostrophecms/login/ui/apos/components/AposForgotPasswordForm.vue
@@ -43,69 +43,12 @@
 </template>
 
 <script>
-import AposLoginFormMixin from 'Modules/@apostrophecms/login/mixins/AposLoginFormMixin';
+import AposForgotPasswordFormLogic from 'Modules/@apostrophecms/login/logic/AposForgotPasswordForm';
 
 export default {
   name: 'AposForgotPasswordForm',
-  mixins: [ AposLoginFormMixin ],
-  emits: [ 'set-stage' ],
-  data() {
-    return {
-      busy: false,
-      done: false,
-      schema: [
-        {
-          name: 'email',
-          label: 'apostrophe:email',
-          placeholder: 'apostrophe:loginEnterEmail',
-          type: 'string',
-          required: true
-        }
-      ]
-    };
-  },
-  computed: {
-    disabled() {
-      return this.doc.hasErrors;
-    },
-    help() {
-      if (this.done) {
-        return this.$t('apostrophe:loginResetRequestDone', {
-          email: this.doc.data.email
-        });
-      }
-      return this.$t('apostrophe:loginResetPasswordRequest');
-    }
-  },
-  created() {
-    if (!this.passwordResetEnabled) {
-      this.$emit('set-stage', 'login');
-    }
-  },
-  methods: {
-    async submit() {
-      if (this.busy) {
-        return;
-      }
-      this.busy = true;
-      this.error = '';
-
-      await this.requestReset();
-    },
-    async requestReset() {
-      try {
-        await apos.http.post(`${apos.login.action}/reset-request`, {
-          busy: true,
-          body: { ...this.doc.data }
-        });
-        this.done = true;
-      } catch (e) {
-        this.error = e.message || 'apostrophe:loginErrorGeneric';
-      } finally {
-        this.busy = false;
-      }
-    }
-  }
+  mixins: [ AposForgotPasswordFormLogic ],
+  emits: [ 'set-stage' ]
 };
 </script>