+++ CHANGELOG.md
@@ -2,19 +2,30 @@
 
 ## UNRELEASED
 
+### Adds
+
+* Adds the possibility to add custom admin bars via the `addBar()` method from the `admin-bar` module.
+
+## 3.49.0 (2023-06-08)
+
 ### Changes
 
-* Updates area UX to not display Add Content controls when a widget is focused
-* Updates area UX to unfocus widget on esc key
-* Updates widget UI to use dashed outlines instead of borders to indicate bounds
-* Updates UI for Insert Menu
-* Updates Insert Menu UX to allow mid-node insertion
+* Updates area UX to not display Add Content controls when a widget is focused.
+* Updates area UX to unfocus widget on esc key.
+* Updates widget UI to use dashed outlines instead of borders to indicate bounds.
+* Updates UI for Insert Menu.
+* Updates Insert Menu UX to allow mid-node insertion.
 * Rich Text Widget's Insert components are now expected to emit `done` and `cancel` for proper RT cleanup. `close` still supported for BC, acts as `done`.
+* Migrated the business logic of the login-related Vue components to external mixins, so that the templates and styles can be overridden by
+copying the component `.vue` file to project level without copying all of the business logic. If you have already copied the components to style them,
+we encourage you to consider replacing your `script` tag with the new version, which just imports the mixin, so that fixes we make there will be
+available in your project.
 
 ### Adds
 
-* Adds keyboard accessibility to Insert menu
+* Adds keyboard accessibility to Insert menu.
 * Adds regex pattern feature for string fields.
+* Adds `pnpm` support. Introduces new optional Apostrophe root configuration `pnpm` to force opt-in/out when auto detection fails. See the [documentation](https://v3.docs.apostrophecms.org/guide/using-pnpm.html) for more details.
 * Adds a warning if database queries involving relationships
 are made before the last `apostrophe:modulesRegistered` handler has fired.
 If you need to call Apostrophe's `find()` methods at startup,
@@ -24,11 +35,20 @@ it is best to wait for the `@apostrophecms/doc:beforeReplicate` event.
 If explicitly set to `false`, prevents the prefix from being automatically added to the URL,  
 when making calls with already-prefixed URLs for instance.
 * Adds the `redirectToFirstLocale` option to the `i18n` module to prevent users from reaching a version of their site that would not match any locale when requesting the site without a locale prefix in the URL.
-* Adds the possibility to add custom admin bars via the `addBar()` method from the `admin-bar` module.
+* If just one instance of a piece type should always exist (per locale if localized), the
+`singletonAuto` option may now be set to `true` or to an object with a `slug` option in
+order to guarantee it. This implicitly sets `singleton: true` as well. This is now used
+internally by `@apostrophecms/global` as well as the optional `@apostrophecms-pro/palette` module.
 
 ### Fixes
 
 * Fix 404 error when viewing/editing a doc which draft has a different version of the slug than the published one.
+* Fixed a bug where multiple home pages can potentially be inserted into the database if the
+default locale is renamed. Introduced the `async apos.doc.bestAposDocId(criteria)` method to
+help identify the right `aposDocId` when inserting a document that might exist in
+other locales.
+* Fixed a bug where singletons like the global doc might not be inserted at all if they
+exist under the former name of the default locale and there are no other locales.
 
 ## 3.48.0 (2023-05-26)